Separate UIConfig into extensions in separate files
