+++ Sources/StreamChatUI/UIConfig+ChannelList.swift
@@ -1,0 +1,38 @@
+//
+// Copyright © 2021 Stream.io Inc. All rights reserved.
+//
+
+import UIKit.UICollectionViewLayout
+
+public extension _UIConfig {
+    struct ChannelListUI {
+        public var channelCollectionView: ChatChannelListCollectionView.Type = ChatChannelListCollectionView.self
+        public var channelCollectionLayout: UICollectionViewLayout.Type = ChatChannelListCollectionViewLayout.self
+        public var channelListSwipeableItemView: _ChatChannelSwipeableListItemView<ExtraData>.Type =
+            _ChatChannelSwipeableListItemView<ExtraData>.self
+        public var channelListItemView: _ChatChannelListItemView<ExtraData>.Type = _ChatChannelListItemView<ExtraData>.self
+        public var channelViewCell: _ChatChannelListCollectionViewCell<ExtraData>.Type =
+            _ChatChannelListCollectionViewCell<ExtraData>.self
+        public var newChannelButton: ChatChannelCreateNewButton<ExtraData>.Type = ChatChannelCreateNewButton<ExtraData>.self
+        public var channelNamer: ChatChannelNamer.Type = ChatChannelNamer.self
+        public var channelListItemSubviews = ChannelListItemSubviews()
+    }
+    
+    struct ChannelListItemSubviews {
+        public var avatarView: _ChatChannelAvatarView<ExtraData>.Type = _ChatChannelAvatarView.self
+        public var unreadCountView: _ChatChannelUnreadCountView<ExtraData>.Type = _ChatChannelUnreadCountView<ExtraData>.self
+        public var readStatusView: _ChatChannelReadStatusCheckmarkView<ExtraData>.Type =
+            _ChatChannelReadStatusCheckmarkView<ExtraData>.self
+        public var onlineIndicator: _ChatOnlineIndicatorView<ExtraData>.Type = _ChatOnlineIndicatorView<ExtraData>.self
+    }
+}
+
+// MARK: - CurrentUser
+
+public extension _UIConfig {
+    struct CurrentUserUI {
+        public var currentUserViewAvatarView: _CurrentChatUserAvatarView<ExtraData>.Type = _CurrentChatUserAvatarView<ExtraData>
+            .self
+        public var avatarView: ChatAvatarView.Type = ChatAvatarView.self
+    }
+}
